+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Moved exceptions from `ogn.exceptions` to `ogn.parser.exceptions`
+- Moved parsing from `ogn.model.*` to `ogn.parser`
 
 ## 0.2.1 - 2016-02-17
 First release via PyPi.